- Improved rendering of fake floors. - Fixed some compilation errors in common.h and enabled using typedefs for vint types. - Added a HAVE_STDINT_H define in blargg_common.h, to allow the use of stdint.h file when compiling. - Fixed a crash when Direct3D renderer updated camera textures. - Used the code analysis tool from MSVC on Vavoom project and fixed some warnings, which seem to improve performance. - Made Vavoom use srand method to seed a value to further randomize the random number generator.

+++ source/p_terrain.cpp
@@ -211,19 +211,19 @@ static void ParseTerrainScript(VScriptParser* sc)
 				{
 					sc->ExpectNumber();
 					SInfo->ChunkXVelMul = sc->Number < 0 ? 0.0 :
-						(1 << sc->Number) / 256.0;
+						float((1 << sc->Number) / 256);
 				}
 				else if (sc->Check("chunkyvelshift"))
 				{
 					sc->ExpectNumber();
 					SInfo->ChunkYVelMul = sc->Number < 0 ? 0.0 :
-						(1 << sc->Number) / 256.0;
+						float((1 << sc->Number) / 256);
 				}
 				else if (sc->Check("chunkzvelshift"))
 				{
 					sc->ExpectNumber();
 					SInfo->ChunkZVelMul = sc->Number < 0 ? 0.0 :
-						(1 << sc->Number) / 256.0;
+						float((1 << sc->Number) / 256);
 				}
 				else if (sc->Check("chunkbasezvel"))
 				{
